refactor: rename EditBlueprint.tsx to BlueprintPlugin.tsx
